refactor: cut out domain as a crate
