+++ domain/src/model.rs
@@ -1,0 +1,14 @@
+pub use coordinate::Coordinate;
+pub use operation::{Operation, OperationType};
+pub use route::{Route, RouteInfo};
+pub use segment::{Segment, SegmentList};
+pub use types::{Distance, Elevation, Latitude, Longitude, Polyline, RouteId};
+
+pub use self::gpx::RouteGpx;
+
+pub(crate) mod coordinate;
+pub(crate) mod gpx;
+pub(crate) mod operation;
+pub(crate) mod route;
+pub(crate) mod segment;
+pub(crate) mod types;
