+++ domain/src/model/gpx.rs
@@ -5,8 +5,9 @@ use std::str::from_utf8;
 use quick_xml::events::Event;
 use quick_xml::{Reader, Writer};
 
-use crate::domain::model::route::Route;
-use crate::utils::error::{ApplicationError, ApplicationResult};
+use route_bucket_utils::{ApplicationError, ApplicationResult};
+
+use crate::model::route::Route;
 
 pub struct RouteGpx(Vec<u8>);
