--- src/domain/model/mod.rs
@@ -1,6 +1,0 @@
-pub mod coordinate;
-pub mod gpx;
-pub mod operation;
-pub mod route;
-pub mod segment;
-pub mod types;
