--- src/domain/service/route.rs
@@ -1,1 +1,0 @@
-
